requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
django-countries==7.5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1     
drf-yasg==1.21.7
gunicorn==21.2.0       
idna==3.6
inflection==0.5.1
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
